move `script` to a top-level module, pull `opcodes` into it
