+++ fuzz/fuzz_targets/uint128_fuzz.rs
@@ -1,6 +1,4 @@
 extern crate bitcoin;
-use std::str::FromStr;
-use std::convert::Into;
 
 fn do_test(data: &[u8]) {
     macro_rules! read_ints {
