+++ src/lib.rs
@@ -61,6 +61,7 @@ mod internal_macros;
 #[macro_use]
 pub mod network;
 pub mod primitives;
+pub mod script;
 pub mod util;
 pub mod consensus;
 pub mod hash_types;
@@ -68,7 +69,6 @@ pub mod hash_types;
 pub use hash_types::*;
 pub use primitives::block::Block;
 pub use primitives::block::BlockHeader;
-pub use primitives::script::Script;
 pub use primitives::transaction::Transaction;
 pub use primitives::transaction::TxIn;
 pub use primitives::transaction::TxOut;
@@ -76,6 +76,7 @@ pub use primitives::transaction::OutPoint;
 pub use primitives::transaction::SigHashType;
 pub use consensus::encode::VarInt;
 pub use network::constants::Network;
+pub use script::Script;
 pub use util::Error;
 pub use util::address::Address;
 pub use util::address::AddressType;
