+++ src/primitives/mod.rs
@@ -19,8 +19,6 @@
 //!
 
 pub mod constants;
-pub mod opcodes;
-pub mod script;
 pub mod transaction;
 pub mod block;
